+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [8.0.1] - 2026-09-16
+## [Unreleased]
 
 ### Fixed